% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeFromDentition.R
\docType{package}
\name{AgeFromDentition-package}
\alias{AgeFromDentition}
\alias{AgeFromDentition-package}
\title{AgeFromDentition: Estimate Dental Age}
\description{
Estimate dental age given a set of dental scores for six teeth following the method of Šešelj M, Sherwood RJ, Konigsberg LW. 2019. Timing of Development of the Permanent Mandibular Dentition: New Reference Values from the Fels Longitudinal Study. Anat Rec 302:1733–1753.
}
\author{
\strong{Maintainer}: Kevin Middleton \email{middletonk@missouri.edu} (\href{https://orcid.org/0000-0003-4704-1064}{ORCID})

Authors:
\itemize{
  \item Kevin Middleton \email{middletonk@missouri.edu} (\href{https://orcid.org/0000-0003-4704-1064}{ORCID})
}

}
\keyword{internal}
