% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dental_age.R
\name{print.dental_age}
\alias{print.dental_age}
\title{Print a dental age estimate}
\usage{
\method{print}{dental_age}(x, ...)
}
\arguments{
\item{x}{a \code{dental_age} object from \code{\link[=estimate_dental_age]{estimate_dental_age()}}.}

\item{...}{ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a dental age estimate
}
\examples{
print(estimate_dental_age(ExampleScores[1, ], verbose = FALSE))

}
